JDL: fixed user management flag passing
Removed the business error checker and split its previous role in two files: one for jdl objects with apps, one without.
Copied the tests from the business error checker and adapted them for the new files.

Fixes #12366
Closes #12439
